+++ velora_project/settings.py
@@ -39,6 +39,7 @@ INSTALLED_APPS = [
 MIDDLEWARE = [
     "django.middleware.security.SecurityMiddleware",
     "django.contrib.sessions.middleware.SessionMiddleware",
+    "django.middleware.locale.LocaleMiddleware",
     "django.middleware.common.CommonMiddleware",
     "django.middleware.csrf.CsrfViewMiddleware",
     "django.contrib.auth.middleware.AuthenticationMiddleware",
@@ -56,6 +57,7 @@ TEMPLATES = [
         "OPTIONS": {
             "context_processors": [
                 "django.template.context_processors.request",
+                "django.template.context_processors.i18n",
                 "django.contrib.auth.context_processors.auth",
                 "django.contrib.messages.context_processors.messages",
                 "velora_ui.context_processors.header_defaults",
@@ -73,10 +75,18 @@ DATABASES = {
     ),
 }
 
-LANGUAGE_CODE = "it-it"
+LANGUAGE_CODE = "it"
 TIME_ZONE = "Europe/Rome"
 USE_I18N = True
 USE_TZ = True
+LANGUAGES = [
+    ("it", "Italiano"),
+    ("en", "English"),
+]
+LOCALE_PATHS = [
+    BASE_DIR / "src" / "velora_ui" / "locale",
+    BASE_DIR / "showcase" / "locale",
+]
 
 STATIC_URL = "static/"
 STATIC_ROOT = BASE_DIR / "staticfiles"