+++ velora_project/urls.py
@@ -1,5 +1,7 @@
 from django.urls import include, path
+from django.views.i18n import set_language
 
 urlpatterns = [
+    path("i18n/setlang/", set_language, name="set_language"),
     path("", include("showcase.urls", namespace="showcase")),
 ]